Modified structure, VAM func, & Readme
